requirements.txt
Pillow==11.3.0
blessed==1.21.0
opencv-python==4.12.0.88
